New hierarchy of classes
By teacher's demand: LiteratureRegister is to hold all objects of Literature.

Objects of Literature's direct subclasses (e.g. Book) only has a name and a publisher. Simple as that.

But! Objects of Literature's abstract subclass "SerialLiterature" (e.g. Magazine) also have a release number, which make them compatible with the Series class, which is also going to be treated as literature down the road. (Or perhaps implemented into the register by an interface?)