Huge cleanup operation
New packages neatly divided by responsibility (and access... "protected" is not as protected as the name suggests).

There is now a messenger (Librarian) doing the heavy lifting between the UI and Literature Register, and a universal standard product type number system for menus and switch cases:
Book = 1,
Magazine = 2,
Journal = 3,
Newspaper = 4.
Currently playing around with a system to automate this numbering for the sake of cohesion.