Made all books into SerialBooks
The old Book has been removed, as has SerialLiterature, as Book was the only class using it.

What was SerialBook has now become Book, this is because having different classes for books in and out of series just causes more problems than it solves.

What was SerialLiterature has become Literature, this is just cosmetic and to make the class structure a bit more logical in relation to the real world.

What was Literature is now Entries. This is purely cosmetic and to aid in the logic of the structure.

All uses of what was Book, have been changed and updated to use the new Book and use its signature.