Table now shows register, and remove button now works
Application can now show the entries in the literature archive (except entries inside a series) in the table, and the remove button now works for items you have selected in the table (alert pop-up if no entry selected). Also copied the fillRegister() method from ApplicationUI and into the ApplicationGUI class for the sake of convenience.