Added View Info button and documentation
Added button to view details about entries in the literature register, with different information being available depending on the class of the entry. If you view info about a series, it will also display a TableView of all entries in the selected series.
Also added documentation for most of the methods in the ApplicationGUI class, and improved cohesion in some places.
Note, TableColumn warnings count now up to 14 instead of 9.